# server/docs/swagger.yaml
basePath: /
definitions:
...
  model.DnsUser:
    properties:
      accessKey:
        type: string
      accessSecret:
        type: string
      account:
        type: integer
      createdAt:
        description: 创建时间
        type: string
      id:
        description: 主键ID
        type: integer
      name:
        type: string
      platform:
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
  model.DomainInfo:
    properties:
      account:
        type: integer
      company:
        type: integer
      createdAt:
        description: 创建时间
        type: string
      dnsProvider:
        type: integer
      domain:
        type: string
      domainId:
        type: string
      id:
        description: 主键ID
        type: integer
      recordTotal:
        type: integer
      registerTime:
        type: string
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
  request.DnsUserSearch:
    properties:
      accessKey:
        type: string
      accessSecret:
        type: string
      account:
        type: integer
      createdAt:
        description: 创建时间
        type: string
      id:
        description: 主键ID
        type: integer
      name:
        type: string
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
      platform:
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
  request.DomainListSearch:
    properties:
      account:
        type: integer
      company:
        type: integer
      createdAt:
        description: 创建时间
        type: string
      dnsProvider:
        type: integer
      domain:
        type: string
      domainId:
        type: string
      id:
        description: 主键ID
        type: integer
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
      recordTotal:
        type: integer
      registerTime:
        type: string
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
  request.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
info:
  contact: {}
  description: This is a sample Server pets
  title: Swagger Example API
  version: 0.0.1
paths:
...
  /dnsUser/createDnsUser:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建DnsUser
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.DnsUser'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建DnsUser
      tags:
      - DnsUser
  /dnsUser/deleteDnsUser:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除DnsUser
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.DnsUser'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除DnsUser
      tags:
      - DnsUser
  /dnsUser/deleteDnsUserByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除DnsUser
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除DnsUser
      tags:
      - DnsUser
  /dnsUser/findDnsUser:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用id查询DnsUser
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.DnsUser'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用id查询DnsUser
      tags:
      - DnsUser
  /dnsUser/getDnsUserList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 分页获取DnsUser列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.DnsUserSearch'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取DnsUser列表
      tags:
      - DnsUser
  /dnsUser/updateDnsUser:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新DnsUser
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.DnsUser'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新DnsUser
      tags:
      - DnsUser
  /domainList/createDomain:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建Domain
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.DomainInfo'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建Domain
      tags:
      - DomainInfo
  /domainList/deleteDomain:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除Domain
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.DomainInfo'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除Domain
      tags:
      - DomainInfo
  /domainList/deleteDomainByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除Domains
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除Domains
      tags:
      - DomainInfo
  /domainList/findDomain:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用id查询Domain
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.DomainInfo'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用id查询Domain
      tags:
      - DomainInfo
  /domainList/flushDomains:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 从线上获取最新域名信息并更新到数据库
      tags:
      - DomainInfo
  /domainList/getDomainList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 分页获取Domain列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.DomainListSearch'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取Domain列表
      tags:
      - DomainInfo
  /domainList/updateDomain:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新Domain
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.DomainInfo'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新Domain
      tags:
      - DomainInfo
  /domainList/updateDomains:
    put:
      consumes:
      - application/json
      parameters:
      - description: 批量更新Domain
        in: body
        name: data
        required: true
        schema:
          items:
            $ref: '#/definitions/model.DomainInfo'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量更新Domain
      tags:
      - DomainInfo
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"